Update unifi_ssl_import.sh
Amending last PR Merge.

The changes added previous streamlined the Let's Encrypt Process but in the process broke the script for user using their own certs. Resulting in errors:
- "No Certificate matches private key"
- "Alias <unifi> does not exist"
etc

Now the script should check if the signed cert exists, if it does add that to the export. (alternative would be to create a combined file, but I personally prefer to keep CA and Cert seperate)

Also added the "deststoretype" option as controller was complaining about security.